결제 API 데이터베이스 붙였으며, transaction 을 이용한 CRUD 작업중입니다.

+++ src/main/java/com/daalgae/daalgaeproject/board/service/BoardService.java
@@ -7,7 +7,6 @@ import com.daalgae.daalgaeproject.common.exception.board.ReplyRegistException;
 import com.daalgae.daalgaeproject.common.exception.board.ReplyRemoveException;
 import com.daalgae.daalgaeproject.common.exception.thumbnail.ThumbnailRegistException;
 import com.daalgae.daalgaeproject.common.paging.SelectCriteria;
-import org.apache.ibatis.annotations.Param;
 
 import java.util.List;
 import java.util.Map;
@@ -21,10 +20,10 @@ public interface BoardService {
     public List<BoardDTO> selectBoardList(SelectCriteria selectCriteria);
 
     /* 게시글 상세 페이지 조회용 메소드 */
-    public BoardDTO selectBoardDetail(int no);
+    public BoardDTO selectBoardDetail(Long no);
 
     /* 해당 게시글의 전체 댓글 조회용 메소드 */
-    public List<ReplyDTO> selectAllReplyList(int boardNo);
+    public List<ReplyDTO> selectAllReplyList(Long boardNo);
 
 
     /* 댓글 등록용 메소드 */
@@ -44,5 +43,5 @@ public interface BoardService {
 
 
     /* 게시글 상세 페이지 조회용 메소드 */
-    public BoardDTO selectThumbnailDetail(int no) ;
+    public BoardDTO selectThumbnailDetail(Long no) ;
 }
